requirements.txt
requests>=2.32.3
colorama>=0.4.6
beautifulsoup4>=4.12.3
lxml>=6.0.0
whois>=1.20240129.2
python-dotenv>=0.21.0
